backend folder/eye_project_backend/requirements.txt
Django>=4.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
firebase-admin
razorpay
requests
gunicorn
whitenoise
python-dateutil
